moved common code to utils
